git-svn-id: file:///semplestb/final_repo@2338 02200ff9-b5b2-46f0-9171-221b09c08c7b

+++ SemplestServices/src/semplest/server/service/springjdbc/storedproc/UpdateRemainingBudgetInCycleSP.java
@@ -20,7 +20,7 @@ public class UpdateRemainingBudgetInCycleSP  extends StoredProcedure
 		declareParameter(new SqlParameter("PromotionPK", Types.INTEGER));
 		declareParameter(new SqlParameter("StartDate", Types.TIMESTAMP));
 		declareParameter(new SqlParameter("EndDate", Types.TIMESTAMP));
-		declareParameter(new SqlOutParameter("ID", Types.BIGINT));
+		declareParameter(new SqlOutParameter("ID", Types.INTEGER));
 		compile();
 	}
 
@@ -36,7 +36,7 @@ public class UpdateRemainingBudgetInCycleSP  extends StoredProcedure
 		}
 		else
 		{
-			return ((Long)results.get("ID")).intValue();
+			return ((Integer)results.get("ID"));
 		}
 	}
 }
